lint: fix or ignore uses of console.log
