+++ packages/core/src/shared/fs/templateRegistry.ts
@@ -16,7 +16,7 @@ import { getLogger } from '../logger'
 import globals from '../extensionGlobals'
 import { Timeout } from '../utilities/timeoutUtils'
 import { localize } from '../utilities/vsCodeUtils'
-import { PerfLog } from '../logger/logger'
+import { PerfLog } from '../logger/perfLogger'
 import { showMessageWithCancel } from '../utilities/messages'
 
 export class CloudFormationTemplateRegistry extends WatchedFiles<CloudFormation.Template> {
