+++ packages/core/src/shared/logger/logger.ts
@@ -156,23 +156,28 @@ export class ConsoleLogger implements Logger {
         }
     }
     public debug(message: string | Error, ...meta: any[]): number {
+        // eslint-disable-next-line aws-toolkits/no-console-log
         console.debug(message, ...meta)
         return 0
     }
     public verbose(message: string | Error, ...meta: any[]): number {
+        // eslint-disable-next-line aws-toolkits/no-console-log
         console.debug(message, ...meta)
         return 0
     }
     public info(message: string | Error, ...meta: any[]): number {
+        // eslint-disable-next-line aws-toolkits/no-console-log
         console.info(message, ...meta)
         return 0
     }
     public warn(message: string | Error, ...meta: any[]): number {
+        // eslint-disable-next-line aws-toolkits/no-console-log
         console.warn(message, ...meta)
         return 0
     }
     /** Note: In nodejs this prints to `stderr` (see {@link Console.error}). */
     public error(message: string | Error, ...meta: any[]): number {
+        // eslint-disable-next-line aws-toolkits/no-console-log
         console.error(message, ...meta)
         return 0
     }
@@ -192,27 +197,4 @@ export function getNullLogger(type?: 'channel' | 'debugConsole' | 'main'): Logge
  */
 export function setLogger(logger: Logger | undefined, type?: 'channel' | 'debugConsole' | 'main') {
     toolkitLoggers[type ?? 'main'] = logger
-}
-
-export class PerfLog {
-    private readonly log
-    public readonly start
-
-    public constructor(public readonly topic: string) {
-        const log = getLogger()
-        this.log = log
-        this.start = performance.now()
-    }
-
-    public elapsed(): number {
-        return performance.now() - this.start
-    }
-
-    public done(): void {
-        if (!this.log.logLevelEnabled('verbose')) {
-            return
-        }
-        const elapsed = this.elapsed()
-        this.log.verbose('%s took %dms', this.topic, elapsed.toFixed(1))
-    }
 }
