+++ packages/core/src/shared/sam/cli/samCliLocator.ts
@@ -9,7 +9,7 @@ import * as filesystemUtilities from '../../filesystemUtilities'
 import { getLogger, Logger } from '../../logger'
 import { SamCliInfoInvocation } from './samCliInfo'
 import { DefaultSamCliValidator, SamCliValidatorContext, SamCliVersionValidation } from './samCliValidator'
-import { PerfLog } from '../../logger/logger'
+import { PerfLog } from '../../logger/perfLogger'
 import { tryRun } from '../../utilities/pathFind'
 
 export class SamCliLocationProvider {
